// frontend/src/app/shared/generated/models/search-result-item-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SearchResultItemDto {
  analysisId: (number | string);
  calibratedValue?: (number | string | null);
  capturedAtUtc?: (string | null);
  completedAtUtc?: (string | null);
  instrumentId?: (number | string | null);
  isLocked: boolean;
  readingId?: (number | string | null);
  readingUnit?: (string | null);
  readingValue?: (number | string | null);
  sampleId: (number | string);
  sampleIdentifier: string;
  site: string;
  startedAtUtc: string;
  status: string;
  templateName: string;
  testId?: (number | string | null);
  validationResult?: (string | null);
}