// frontend/src/app/shared/generated/models/paged-result-of-search-result-item-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SearchResultItemDto } from '../models/search-result-item-dto';
export interface PagedResultOfSearchResultItemDto {
  items: Array<SearchResultItemDto>;
  pageNumber: (number | string);
  pageSize: (number | string);
  totalCount: (number | string);
}